refactor(core): inline variance classification into css-axis-projected
