+++ packages/core/src/css-axis-projected.ts
@@ -3,10 +3,10 @@ import { generateShorthand, makeCSSVar, transformCSSValue } from '@terrazzo/toke
 import { CHROME_ROLES, CHROME_VAR_PREFIX, DEFAULT_CHROME_MAP } from '#/chrome.ts';
 import { cssEscape } from '#/css-escape.ts';
 import { dataAttr } from '#/data-attr.ts';
+import { listPaths } from '#/token-graph/queries.ts';
 import { resolveAliasAllAt, resolveAllAt } from '#/token-graph/walk.ts';
 import type { Project, TokenMap } from '#/types.ts';
-import { analyzeProjectVariance } from '#/variance-analysis.ts';
-import type { VarianceInfo } from '#/variance-analysis.ts';
+import { permutationID } from '#/types.ts';
 import { valueKey } from '#/value-key.ts';
 
 /**
@@ -22,6 +22,115 @@ type RawTokenMap = Record<string, TokenNormalized>;
 
 function asRawTokens(map: TokenMap): RawTokenMap {
   return map as RawTokenMap;
+}
+
+/** Distinguishes a token's variance shape so the emitter can route it. */
+type VarianceInfo =
+  | {
+      kind: 'baseline-only';
+    }
+  | {
+      kind: 'single-axis';
+      axis: string;
+    }
+  | {
+      kind: 'orthogonal-after-probe';
+      touching: ReadonlySet<string>;
+    }
+  | {
+      kind: 'joint-variant';
+      touching: ReadonlySet<string>;
+      jointCases: readonly JointCase[];
+    };
+
+/** One concrete `(A=ctx_a, B=ctx_b)` combination where projection differs from cartesian. */
+interface JointCase {
+  axisA: string;
+  ctxA: string;
+  axisB: string;
+  ctxB: string;
+  cartesianValueKey: string;
+  permutationName: string;
+}
+
+/**
+ * Classify every token in a project. Returns a Map keyed by token path.
+ * Reads from `project.tokenGraph` — the sole resolution surface.
+ */
+function analyzeProjectVariance(project: Project): Map<string, VarianceInfo> {
+  const result = new Map<string, VarianceInfo>();
+  const { axes, tokenGraph, defaultTuple } = project;
+
+  if (axes.length === 0) {
+    for (const path of listPaths(tokenGraph)) {
+      result.set(path, { kind: 'baseline-only' });
+    }
+    return result;
+  }
+
+  for (const path of listPaths(tokenGraph)) {
+    const node = tokenGraph.nodes[path];
+    if (!node) continue;
+    const touching: Set<string> = new Set(node.affectedBy);
+
+    if (touching.size === 0) {
+      result.set(path, { kind: 'baseline-only' });
+      continue;
+    }
+
+    if (touching.size === 1) {
+      const [axis] = touching;
+      if (axis === undefined) continue;
+      result.set(path, { kind: 'single-axis', axis });
+      continue;
+    }
+
+    const touchingAxes = [...touching];
+    const jointCases: JointCase[] = [];
+
+    for (let i = 0; i < touchingAxes.length; i++) {
+      for (let j = i + 1; j < touchingAxes.length; j++) {
+        const axisA = touchingAxes[i]!;
+        const axisB = touchingAxes[j]!;
+        const axisADef = tokenGraph.axisDefaults[axisA] ?? '';
+        const axisBDef = tokenGraph.axisDefaults[axisB] ?? '';
+        const axisACtxs = (tokenGraph.axisContexts[axisA] ?? []).filter((c) => c !== axisADef);
+        const axisBCtxs = (tokenGraph.axisContexts[axisB] ?? []).filter((c) => c !== axisBDef);
+
+        for (const ctxA of axisACtxs) {
+          for (const ctxB of axisBCtxs) {
+            const jointTuple = { ...defaultTuple, [axisA]: ctxA, [axisB]: ctxB };
+            const cartesianToken = resolveAllAt(tokenGraph, jointTuple)[path];
+            const cartKey = valueKey(cartesianToken);
+
+            const afterA = resolveAllAt(tokenGraph, { ...defaultTuple, [axisA]: ctxA })[path];
+            const afterB = resolveAllAt(tokenGraph, { ...defaultTuple, [axisB]: ctxB })[path];
+            const projKeyAB = valueKey(afterB);
+            const projKeyBA = valueKey(afterA);
+
+            if (cartKey !== projKeyAB && cartKey !== projKeyBA) {
+              jointCases.push({
+                axisA,
+                ctxA,
+                axisB,
+                ctxB,
+                cartesianValueKey: cartKey,
+                permutationName: permutationID(jointTuple),
+              });
+            }
+          }
+        }
+      }
+    }
+
+    if (jointCases.length === 0) {
+      result.set(path, { kind: 'orthogonal-after-probe', touching });
+    } else {
+      result.set(path, { kind: 'joint-variant', touching, jointCases });
+    }
+  }
+
+  return result;
 }
 
 /** @internal Addon-internal smart-emitter options. Not part of the public API. */
